+++ drivers/gpu/drm/apple/Makefile
@@ -5,6 +5,8 @@ CFLAGS_trace.o = -I$(src)
 appledrm-y := apple_drv.o
 
 apple_dcp-y := dcp.o dcp_backlight.o iomfb.o parser.o
+apple_dcp-y += iomfb_v12_3.o
+apple_dcp-y += iomfb_v13_2.o
 apple_dcp-$(CONFIG_TRACING) += trace.o
 
 apple_piodma-y := dummy-piodma.o
