+++ drivers/gpu/drm/apple/iomfb.h
@@ -6,6 +6,8 @@
 
 #include <linux/types.h>
 
+#include "version_utils.h"
+
 /* Fixed size of shared memory between DCP and AP */
 #define DCP_SHMEM_SIZE 0x100000
 
@@ -106,35 +108,6 @@ struct dcp_rect {
  */
 #define IOMFB_SET_BACKGROUND	BIT(31)
 
-struct dcp_swap {
-	u64 ts1;
-	u64 ts2;
-	u64 unk_10[6];
-	u64 flags1;
-	u64 flags2;
-
-	u32 swap_id;
-
-	u32 surf_ids[SWAP_SURFACES];
-	struct dcp_rect src_rect[SWAP_SURFACES];
-	u32 surf_flags[SWAP_SURFACES];
-	u32 surf_unk[SWAP_SURFACES];
-	struct dcp_rect dst_rect[SWAP_SURFACES];
-	u32 swap_enabled;
-	u32 swap_completed;
-
-	u32 bg_color;
-	u8 unk_110[0x1b8];
-	u32 unk_2c8;
-	u8 unk_2cc[0x14];
-	u32 unk_2e0;
-	u16 unk_2e2;
-	u64 bl_unk;
-	u32 bl_value; // min value is 0x10000000
-	u8  bl_power; // constant 0x40 for on
-	u8 unk_2f3[0x2d];
-} __packed;
-
 /* Information describing a plane of a planar compressed surface */
 struct dcp_plane_info {
 	u32 width;
@@ -152,38 +125,6 @@ struct dcp_plane_info {
 struct dcp_component_types {
 	u8 count;
 	u8 types[7];
-} __packed;
-
-/* Information describing a surface */
-struct dcp_surface {
-	u8 is_tiled;
-	u8 is_tearing_allowed;
-	u8 is_premultiplied;
-	u32 plane_cnt;
-	u32 plane_cnt2;
-	u32 format; /* DCP fourcc */
-	u32 ycbcr_matrix;
-	u8 xfer_func;
-	u8 colorspace;
-	u32 stride;
-	u16 pix_size;
-	u8 pel_w;
-	u8 pel_h;
-	u32 offset;
-	u32 width;
-	u32 height;
-	u32 buf_size;
-	u64 protection_opts;
-	u32 surface_id;
-	struct dcp_component_types comp_types[MAX_PLANES];
-	u64 has_comp;
-	struct dcp_plane_info planes[MAX_PLANES];
-	u64 has_planes;
-	u32 compression_info[MAX_PLANES][13];
-	u64 has_compr_info;
-	u32 unk_num;
-	u32 unk_denom;
-	u8 padding[7];
 } __packed;
 
 struct dcp_rt_bandwidth {
@@ -218,13 +159,21 @@ enum dcpep_method {
 	iomfbep_a132_backlight_service_matched,
 	iomfbep_a358_vi_set_temperature_hint,
 	iomfbep_get_color_remap_mode,
+	iomfbep_last_client_close,
 	dcpep_num_methods
 };
+
+#define IOMFB_METHOD(tag, name) [name] = { #name, tag }
 
 struct dcp_method_entry {
 	const char *name;
 	char tag[4];
 };
+
+#define IOMFB_MAX_CB (1000)
+struct apple_dcp;
+
+typedef bool (*iomfb_cb_handler)(struct apple_dcp *, int, void *, void *);
 
 /* Prototypes */
 
@@ -287,21 +236,6 @@ struct dcp_map_physical_resp {
 	u32 mem_desc_id;
 } __packed;
 
-struct dcp_map_reg_req {
-	char obj[4];
-	u32 index;
-	u32 flags;
-	u8 addr_null;
-	u8 length_null;
-	u8 padding[2];
-} __packed;
-
-struct dcp_map_reg_resp {
-	u64 addr;
-	u64 length;
-	u32 ret;
-} __packed;
-
 struct dcp_swap_start_req {
 	u32 swap_id;
 	struct dcp_iouserclient client;
@@ -314,34 +248,6 @@ struct dcp_swap_start_resp {
 	u32 swap_id;
 	struct dcp_iouserclient client;
 	u32 ret;
-} __packed;
-
-struct dcp_swap_submit_req {
-	struct dcp_swap swap;
-	struct dcp_surface surf[SWAP_SURFACES];
-	u64 surf_iova[SWAP_SURFACES];
-	u8 unkbool;
-	u64 unkdouble;
-	u32 clear; // or maybe switch to default fb?
-	u8 swap_null;
-	u8 surf_null[SWAP_SURFACES];
-	u8 unkoutbool_null;
-	u8 padding[1];
-} __packed;
-
-struct dcp_swap_submit_resp {
-	u8 unkoutbool;
-	u32 ret;
-	u8 padding[3];
-} __packed;
-
-struct dc_swap_complete_resp {
-	u32 swap_id;
-	u8 unkbool;
-	u64 swap_data;
-	u8 swap_info[0x6c4];
-	u32 unkint;
-	u8 swap_info_null;
 } __packed;
 
 struct dcp_get_uint_prop_req {
@@ -433,6 +339,15 @@ struct iomfb_get_color_remap_mode_req {
 struct iomfb_get_color_remap_mode_resp {
 	u32 mode;
 	u32 ret;
+} __packed;
+
+struct iomfb_last_client_close_req {
+	u8 unkint_null;
+	u8 padding[3];
+} __packed;
+
+struct iomfb_last_client_close_resp {
+	u32 unkint;
 } __packed;
 
 #endif
