+++ drivers/gpu/drm/apple/iomfb_v13_2.h
@@ -1,0 +1,17 @@
+// SPDX-License-Identifier: GPL-2.0-only OR MIT
+/* Copyright The Asahi Linux Contributors */
+
+#ifndef __APPLE_IOMFB_V13_2_H__
+#define __APPLE_IOMFB_V13_2_H__
+
+#include "version_utils.h"
+
+#define DCP_FW v13_2
+#define DCP_FW_VER DCP_FW_VERSION(13, 2, 0)
+
+#include "iomfb_template.h"
+
+#undef DCP_FW_VER
+#undef DCP_FW
+
+#endif /* __APPLE_IOMFB_V13_2_H__ */
